# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Go untuk manga library: scan/move folder, bookmark, translate, bug report, duplikat.\nLogin lewat POST /login, lalu klik Authorize dan isi `Bearer \u003caccess_token\u003e`.",
        "title": "Manga Library API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/translate/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "List antrian translate (pending/processing/failed)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PendingTranslationItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/translate/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Ambil setting translator (tanpa API key)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/dto.TranslatorSettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Simpan setting translator",
                "parameters": [
                    {
                        "description": "Setting",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TranslatorSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Reset setting translator ke default",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/translate/settings/worker": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Ambil setting translator untuk worker (termasuk API key)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/translate/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Keluarkan manga dari antrian translate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "new_folders.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "409": {
                        "description": "Sedang diproses",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/translate/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Simpan halaman hasil translate sebagai folder baru di DST_DIR dan tandai request completed.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Upload hasil translate (dipakai worker)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "new_folders.id asal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nama folder hasil",
                        "name": "new_folder_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Halaman hasil translate",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "object",
                                            "properties": {
                                                "files_skipped": {
                                                    "type": "integer"
                                                },
                                                "files_written": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "new_folder_id": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/translate/{id}/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Masukkan manga ke antrian translate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "new_folders.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Model.BaseResponseModel"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "type": "object",
                                            "properties": {
                                                "status": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
        "/translate/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translate"
                ],
                "summary": "Update status translate (dipakai worker)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "new_folders.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status: pending|processing|completed|failed",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTranslationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Model.BaseResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Model.BaseResponseModel": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Data": {},
                "HeaderMessage": {
                    "type": "string"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
        "dto.PendingTranslationItem": {
            "type": "object",
            "properties": {
                "folder_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
...
        "dto.TranslatorSettingsRequest": {
            "type": "object",
            "required": [
                "detection_size",
                "detector",
                "gpu_mode",
                "inpainter",
                "inpainting_precision",
                "inpainting_size",
                "ocr",
                "target_lang",
                "translator"
            ],
            "properties": {
                "api_base": {
                    "type": "string"
                },
                "api_key": {
                    "description": "empty = keep the stored key",
                    "type": "string"
                },
                "api_model": {
                    "type": "string"
                },
                "clear_api_key": {
                    "description": "true = remove the stored key",
                    "type": "boolean"
                },
                "detection_size": {
                    "type": "integer"
                },
                "detector": {
                    "type": "string"
                },
                "gpu_mode": {
                    "type": "string"
                },
                "inpainter": {
                    "type": "string"
                },
                "inpainting_precision": {
                    "type": "string"
                },
                "inpainting_size": {
                    "type": "integer"
                },
                "ocr": {
                    "type": "string"
                },
                "target_lang": {
                    "type": "string"
                },
                "translator": {
                    "type": "string"
                }
            }
        },
        "dto.TranslatorSettingsResponse": {
            "type": "object",
            "properties": {
                "api_base": {
                    "type": "string"
                },
                "api_model": {
                    "type": "string"
                },
                "detection_size": {
                    "type": "integer"
                },
                "detector": {
                    "type": "string"
                },
                "gpu_mode": {
                    "type": "string"
                },
                "has_api_key": {
                    "type": "boolean"
                },
                "inpainter": {
                    "type": "string"
                },
                "inpainting_precision": {
                    "type": "string"
                },
                "inpainting_size": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "ocr": {
                    "type": "string"
                },
                "target_lang": {
                    "type": "string"
                },
                "translator": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateTranslationStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "error_message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: `Bearer \u003caccess_token\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}